+++ Realm.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:src/RealmJS.xcodeproj">
-   </FileRef>
-   <FileRef
       location = "group:react-native/ios/RealmReact.xcodeproj">
    </FileRef>
    <FileRef
