+++ cmake/ios.toolchain.cmake
@@ -1,0 +1,5 @@
+include("${CMAKE_CURRENT_LIST_DIR}/../vendor/realm-core/tools/cmake/ios.toolchain.cmake")
+
+set(CMAKE_SYSTEM_NAME iOS)
+set(CMAKE_C_VISIBILITY_PRESET hidden)
+set(CMAKE_CXX_VISIBILITY_PRESET hidden)
