+++ react-native/ios/.gitignore
@@ -1,2 +1,4 @@
 Pods
 RealmReact.xcworkspace
+
+build
