+++ react-native/ios/RealmReact/RealmReact.mm
@@ -19,11 +19,7 @@
 #import "RealmReact.h"
 #import "RealmAnalytics.h"
 
-#import "jsc/jsc_init.h"
-
-#import "impl/realm_coordinator.hpp"
-#import "shared_realm.hpp"
-#import "sync/app.hpp"
+#import <realm-js-ios/jsc_init.h>
 
 #import <React/RCTBridge+Private.h>
 #import <React/RCTJavaScriptExecutor.h>
@@ -35,11 +31,11 @@
 #import <net/if.h>
 
 #if DEBUG
+#include <realm-js-ios/rpc.hpp>
 #import "GCDWebServer.h"
 #import "GCDWebServerDataRequest.h"
 #import "GCDWebServerDataResponse.h"
 #import "GCDWebServerErrorResponse.h"
-#import "rpc.hpp"
 
 #define WEB_SERVER_PORT 8083
 
@@ -73,7 +69,7 @@ extern "C" JSGlobalContextRef RealmReactGetJSGlobalContextForExecutor(id executo
         }
         else {
             // for RN < 0.28.0
-            assert([RCTJavaScriptContext instancesRespondToSelector:@selector(initWithJSContext:)]);
+            NSCAssert([RCTJavaScriptContext instancesRespondToSelector:@selector(initWithJSContext:)], @"React Native version too old");
             rctJSContext = [[RCTJavaScriptContext alloc] initWithJSContext:[JSContext new]];
         }
 
@@ -232,8 +228,8 @@ RCT_REMAP_METHOD(emit, emitEvent:(NSString *)eventName withObject:(id)object) {
             NSData *responseData;
 
             if (rpcServer) {
-                json args = json::parse([[(GCDWebServerDataRequest *)request text] UTF8String]);
-                std::string responseText = rpcServer->perform_request(request.path.UTF8String, std::move(args)).dump();
+                std::string args = [[(GCDWebServerDataRequest *)request text] UTF8String];
+                std::string responseText = rpcServer->perform_request(request.path.UTF8String, args);
 
                 responseData = [NSData dataWithBytes:responseText.c_str() length:responseText.length()];
             }
@@ -268,10 +264,7 @@ RCT_REMAP_METHOD(emit, emitEvent:(NSString *)eventName withObject:(id)object) {
 #endif
 
 - (void)invalidate {
-    // Close all cached Realms
-    realm::_impl::RealmCoordinator::clear_all_caches();
-    // Clear the Object Store App cache, to prevent instances from using a context that was released
-    realm::app::App::clear_cached_apps();
+    RJSInvalidateCaches();
 #if DEBUG
     // shutdown rpc if in chrome debug mode
     [self shutdownRPC];
