+++ src/jsc/rpc.hpp
@@ -18,93 +18,21 @@
 
 #pragma once
 
-#include <functional>
-#include <future>
-#include <thread>
-
-#include "jsc_init.hpp"
-#include "concurrent_deque.hpp"
-#include <external/json/json.hpp>
-#include "jsc_types.hpp"
-#include "jsc_protected.hpp"
-#include "js_network_transport.hpp"
+#include <string>
+#include <memory>
 
 namespace realm {
-
-class ObjectSchema;
-
 namespace rpc {
 
-using json = nlohmann::json;
-
-using RPCObjectID = u_int64_t;
-using RPCRequest = std::function<json(const json)>;
-using NetworkTransportFactory = typename js::JavaScriptNetworkTransport<jsc::Types>::NetworkTransportFactory;
-
-class RPCWorker {
-  public:
-    RPCWorker();
-    ~RPCWorker();
-
-    template<typename Fn>
-    json add_task(Fn&&);
-    void invoke_callback(json);
-    json resolve_callback(json args);
-    std::future<json> add_promise();
-
-    bool try_run_task();
-    void stop();
-    json try_pop_callback();
-    bool should_stop();
-
-  private:
-    bool m_stop = false;
-    int m_depth = 0;
-#if __APPLE__
-    std::thread m_thread;
-    CFRunLoopRef m_loop;
-#endif
-    ConcurrentDeque<std::function<void()>> m_tasks;
-    ConcurrentDeque<std::promise<json>> m_promises;
-    ConcurrentDeque<json> m_callbacks;
-};
-
+class RPCServerImpl;
 class RPCServer {
   public:
     RPCServer();
     ~RPCServer();
-    json perform_request(std::string const& name, json&& args);
+    std::string perform_request(std::string const& name, std::string const& json_args);
     bool try_run_task();
-
-
   private:
-    JSGlobalContextRef m_context;
-    std::mutex m_request_mutex;
-    std::map<std::string, RPCRequest> m_requests;
-    std::map<RPCObjectID, js::Protected<JSObjectRef>> m_objects;
-    std::map<RPCObjectID, js::Protected<JSObjectRef>> m_callbacks;
-    // The key here is the same as the value in m_callbacks. We use the raw pointer as a key here,
-    // because protecting the value in m_callbacks pins the function object and prevents it from being moved
-    // by the garbage collector upon compaction.
-    std::map<JSObjectRef, RPCObjectID> m_callback_ids;
-    RPCObjectID m_session_id;
-    RPCWorker m_worker;
-    u_int64_t m_callback_call_counter;
-    uint64_t m_reset_counter = 0;
-
-    std::mutex m_pending_callbacks_mutex;
-    std::map<std::pair<uint64_t, uint64_t>, std::promise<json>> m_pending_callbacks;
-
-    NetworkTransportFactory previous_transport_generator;
-
-    static JSValueRef run_callback(JSContextRef, JSObjectRef, JSObjectRef, size_t, const JSValueRef[], JSValueRef *exception);
-
-    RPCObjectID store_object(JSObjectRef object);
-    JSObjectRef get_object(RPCObjectID) const;
-    JSObjectRef get_realm_constructor() const;
-
-    json serialize_json_value(JSValueRef value);
-    JSValueRef deserialize_json_value(const json dict);
+    std::unique_ptr<RPCServerImpl> m_impl;
 };
 
 } // rpc
